backend/requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.15
django-appconf==1.0.5
django-cleanup==8.0.0
django-cors-headers==3.14.0
django-crontab==0.7.1
django-imagekit==4.1.0
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.3.0
drf-nested-routers==0.93.4
flit_core==3.6.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
JPype1==1.4.1
MarkupSafe==2.1.2
mysqlclient==1.4.6
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.11
pilkit==2.0
Pillow==9.5.0
pip==22.3.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.7.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.7
requests==2.30.0
requests-oauthlib==1.3.1
setuptools==67.8.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==2.0.2
wheel==0.38.4
wincertstore==0.2
zipp==3.15.0
